@@ -2,7 +2,11 @@
   import github from '$lib/images/github.svg';
   import linkedin from '$lib/images/linkedin.png';
 
-  export let date: string;
+  type FooterProps = {
+    date: string,
+  };
+
+  let { date }: FooterProps = $props();
 </script>
 
 <footer class="flex flex-col justify-evenly items-center mt-8 py-6 gap-6 text-center bg-primary text-white">
