@@ -2,7 +2,7 @@
   import Card from '$lib/components/Card.svelte';
   import PostList from '$lib/components/PostList.svelte';
 
-  export let data;
+  let { data } = $props();
 </script>
 
 <svelte:head>
