@@ -1,9 +1,8 @@
-import type { Project } from '$lib/types';
+import { getProjects } from '$lib/content';
 
 /** @type {import('./$types').PageLoad} */
-export async function load({ fetch }) {
-  const response = await fetch('api/projects');
-  const projects: Project[] = await response.json();
+export async function load() {
+  const projects = await getProjects();
 
   return {
     projects,
